+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds support for placeholders in MOTD messages.
+
 ## [v8.9.4](https://github.com/lexicalunit/spellbot/releases/tag/v8.9.4) - 2023-04-19
 
 ### Changed